requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
altair==4.1.0             # via streamlit
appdirs==1.4.4            # via virtualenv
appnope==0.1.0            # via ipykernel, ipython
argon2-cffi==20.1.0       # via notebook
astor==0.8.1              # via streamlit
async-generator==1.10     # via nbclient
attrs==19.3.0             # via jsonschema, pytest
babel==2.8.0              # via time-window
backcall==0.2.0           # via ipython
base58==2.0.1             # via streamlit
bleach==3.2.1             # via nbconvert
blinker==1.4              # via streamlit
boto3==1.16.13            # via streamlit
botocore==1.19.13         # via boto3, s3transfer, streamlit
cachetools==4.1.1         # via streamlit
certifi==2020.6.20        # via matplotlib, requests
cffi==1.14.3              # via argon2-cffi
chardet==3.0.4            # via requests
click==7.1.2              # via -r requirements.in, pip-tools, streamlit
cycler==0.10.0            # via matplotlib
cython==0.29.17           # via pmdarima
daiquiri==2.1             # via -r requirements.in
decorator==4.4.2          # via ipython, validators
defusedxml==0.6.0         # via nbconvert
distlib==0.3.1            # via virtualenv
entrypoints==0.3          # via altair, nbconvert
enum-compat==0.0.3        # via streamlit
filelock==3.0.12          # via tox, virtualenv
gitdb==4.0.5              # via gitpython
gitpython==3.1.11         # via streamlit
idna==2.10                # via requests
importlib-metadata==1.7.0  # via jsonschema, pluggy, pytest, tox, virtualenv
iniconfig==1.0.1          # via pytest
ipykernel==5.3.4          # via ipywidgets, notebook, pydeck
ipython-genutils==0.2.0   # via nbformat, notebook, traitlets
ipython==7.19.0           # via ipykernel, ipywidgets
ipywidgets==7.5.1         # via pydeck
jedi==0.17.2              # via ipython
jinja2==2.11.2            # via altair, nbconvert, notebook, pydeck
jmespath==0.10.0          # via boto3, botocore
joblib==0.17.0            # via pmdarima, scikit-learn
jsonschema==3.2.0         # via altair, nbformat
jupyter-client==6.1.7     # via ipykernel, nbclient, notebook
jupyter-core==4.6.3       # via jupyter-client, nbconvert, nbformat, notebook
jupyterlab-pygments==0.1.2  # via nbconvert
kiwisolver==1.2.0         # via matplotlib
markupsafe==1.1.1         # via jinja2
matplotlib==3.3.2         # via -r requirements.in, seaborn
mistune==0.8.4            # via nbconvert
more-itertools==8.4.0     # via pytest
mypy-extensions==0.4.3    # via typing-inspect
nbclient==0.5.1           # via nbconvert
nbconvert==6.0.7          # via notebook
nbformat==5.0.8           # via ipywidgets, nbclient, nbconvert, notebook
nest-asyncio==1.4.2       # via nbclient
notebook==6.1.5           # via widgetsnbextension
numpy==1.19.1             # via -r requirements.in, altair, matplotlib, pandas, pandera, patsy, pmdarima, pyarrow, pydeck, ruptures, scikit-learn, scipy, seaborn, statsmodels, streamlit
packaging==20.4           # via bleach, pytest, streamlit, tox
pandas==1.1.4             # via -r requirements.in, altair, pandera, pmdarima, seaborn, statsmodels, streamlit
pandera==0.5.0            # via -r requirements.in
pandocfilters==1.4.3      # via nbconvert
parso==0.7.1              # via jedi
pathtools==0.1.2          # via watchdog
patsy==0.5.1              # via statsmodels
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==7.2.0             # via matplotlib, streamlit
pip-tools==5.3.1          # via -r requirements.in
plotly==4.12.0            # via -r requirements.in
pluggy==0.13.1            # via pytest, tox
pmdarima==1.7.1           # via -r requirements.in
prometheus-client==0.8.0  # via notebook
prompt-toolkit==3.0.8     # via ipython
protobuf==3.13.0          # via streamlit
ptyprocess==0.6.0         # via pexpect, terminado
py==1.9.0                 # via pytest, tox
pyarrow==2.0.0            # via streamlit
pycparser==2.20           # via cffi
pydeck==0.5.0             # via streamlit
pygments==2.7.2           # via ipython, jupyterlab-pygments, nbconvert
pyparsing==2.4.7          # via matplotlib, packaging
pyrsistent==0.17.3        # via jsonschema
pytest-mock==3.3.1        # via -r requirements.in
pytest==6.0               # via -r requirements.in, pytest-mock
python-dateutil==2.8.1    # via botocore, jupyter-client, matplotlib, pandas, streamlit, time-window
python-json-logger==0.1.11  # via daiquiri
pytz==2020.1              # via babel, pandas, tzlocal
pyzmq==19.0.2             # via jupyter-client, notebook
requests==2.24.0          # via streamlit
retrying==1.3.3           # via plotly
ruptures==1.0.5           # via -r requirements.in
s3transfer==0.3.3         # via boto3
scikit-learn==0.23.2      # via -r requirements.in, pmdarima
scipy==1.5.2              # via -r requirements.in, pmdarima, ruptures, scikit-learn, seaborn, statsmodels
seaborn==0.11.0           # via -r requirements.in
send2trash==1.5.0         # via notebook
six==1.15.0               # via argon2-cffi, bleach, cycler, jsonschema, packaging, patsy, pip-tools, plotly, protobuf, python-dateutil, retrying, tox, validators, virtualenv
smmap==3.0.4              # via gitdb
statsmodels==0.11.1       # via pmdarima
streamlit==0.70.0         # via -r requirements.in
terminado==0.9.1          # via notebook
testpath==0.4.4           # via nbconvert
threadpoolctl==2.1.0      # via scikit-learn
time-window==0.1.0        # via -r requirements.in
toml==0.10.1              # via pytest, streamlit, tox
toolz==0.11.1             # via altair
tornado==6.1              # via ipykernel, jupyter-client, notebook, streamlit, terminado
tox==3.14.6               # via -r requirements.in
traitlets==5.0.5          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbclient, nbconvert, nbformat, notebook, pydeck
typing-extensions==3.7.4.3  # via pandera, typing-inspect
typing-inspect==0.6.0     # via pandera
tzlocal==2.1              # via streamlit
urllib3==1.25.11          # via botocore, pmdarima, requests
validators==0.18.1        # via streamlit
virtualenv==20.0.33       # via tox
watchdog==0.10.3          # via streamlit
wcwidth==0.2.5            # via prompt-toolkit
webencodings==0.5.1       # via bleach
widgetsnbextension==3.5.1  # via ipywidgets
wrapt==1.12.1             # via pandera
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
